Integration test for bres upload
Create integration tests to test an upload of a bres sample file is
successful. The integration test requires the following services to be
running:
* postgres
* redis
* rabbitmq
* sftp
* ras-party

The tests could be extended to ensure messages are put on rabbitmq or
data has been inserted into the database. Ideally you would test these
via an API even if that means introducing them to make testing easier.