Using optional dockerfile buildArgs
We have been fighting with the `COPY target/samplesvc*.jar` syntax when
there is  more that one samplesvc*.jar file that  exists in the target
directory. Thought it was fixed by the docker info change but the same
issue occurs when doing a release because it builds the samplesvc-sources.jar.
This change makes the dockerfile JAR_FILE arg optional to allow the maven
build to provide the full name including version.