Script for uploading social sample file
We can't upload social sample files via the UI at the moment because the
type of social file is hard-coded in response-operations-ui. This script
allows a sample to be uploaded in a simple way without the UI.

A simple script to upload social sample files. It will validate the
arguments and then make a curl request to upload a file.