Script for uploading social sample file (#45)
# Background
We can't upload social sample files via the UI at the moment because the
type of social file is hard-coded in response-operations-ui. This script
allows a sample to be uploaded in a simple way without the UI.

A simple script to upload social sample files. It will validate the
arguments and then make a curl request to upload a file.

# How to test
1. Checkout rm-sample-service -b upload-a-social-sample-file
1. Start docker-compose file `docker-compose up -d`
1. Change `application.yml` sections to the postgres user
    ```
    liquibase:
      user: postgres
      password: postgres

    spring:
      datasource:
        username: postgres
        password: postgres
    ```
1. Start service by running `SampleSvcApplication` in intellij
1. Get a sample file `wget https://raw.githubusercontent.com/ONSdigital/rm-sample-service/load-a-social-sample-file/src/test/resources/csv/social-survey-sample.csv`
1. Run script `./scripts/load_social_sample_file.sh  -h http://localhost:8125 -u admin -p secret -f social-survey-sample.csv`