Load a social sample file (#44)
Social files currently have four required attribute fields and can contain an arbitrary number of non mandatory attributes. The attributes are described by a header row in the sample file.

Adds functionality to parse a social sample file, persist sample attributes and post attributes to collection exercise. When a file is uploaded it is parsed and validated. If valid, it will be saved to a sample attributes table in the sample database linked by the sample unit id (UUID). When a request is made for samples the sample unit and sample attributes and put on the sample delivery queue.

We have added integration tests to exercise the upload functionality and ensure they are published to the queue with the sample attributes. 

**How to test:**
Requires changes in these branches:
https://github.com/ONSdigital/rm-samplesvc-api/pull/11
https://github.com/ONSdigital/rm-party-service-api/pull/11

You'll have to build them locally then point the pom file to those snapshots, this will block Travis from building until those changes are on master.

Try loading a social sample through the endpoint using the loader script:
https://github.com/ONSdigital/rm-sample-service/pull/45

You should also try the loader script with a full example sample file, there is one available as an attachment on the story card:
![](https://github.trello.services/images/mini-trello-icon.png) [Dev task 1: INT: load a social sample file (13)](https://trello.com/c/lTUmPnuf/97-dev-task-1-int-load-a-social-sample-file-13)

Also requires these changes before being merged:
https://github.com/ONSdigital/rm-samplesvc-api/pull/9
https://github.com/ONSdigital/rm-sample-service/pull/42